requirements.txt
Flask==2.1.2
flask-cors==3.0.10
tensorflow==2.8.0
Pillow==9.0.1
numpy==1.22.3
